Move "StatusCard" component to pages/Jobs/Runs folder
