+++ operator_ui/src/pages/Jobs/Runs/Show.tsx
@@ -4,7 +4,7 @@ import { Route, RouteComponentProps, Switch } from 'react-router-dom'
 import { CardTitle } from '@chainlink/styleguide'
 import { Card, Grid, Typography } from '@material-ui/core'
 import Content from 'components/Content'
-import StatusCard from 'components/StatusCard/StatusCard'
+import StatusCard from './StatusCard'
 import { useErrorHandler } from 'hooks/useErrorHandler'
 import { useLoadingPlaceholder } from 'hooks/useLoadingPlaceholder'
 import RegionalNav from './RegionalNav'
